+++ frontend/javascripts/dashboard/dataset/dataset_import_view.js
@@ -499,23 +499,28 @@ class DatasetImportView extends React.PureComponent<Props, State> {
                   </Hideable>
                 </TabPane>
 
-                <TabPane tab={<span>Sharing</span>} key="sharing" forceRender>
-                  <Hideable hidden={this.state.activeTabKey !== "sharing"}>
-                    <ImportSharingComponent form={form} datasetId={this.props.datasetId} />
-                  </Hideable>
-                </TabPane>
-
                 <TabPane
                   tab={
                     <span>
-                      Metadata {formErrors.general ? errorIcon : hasNoAllowedTeamsWarning}
+                      Sharing & Permissions{" "}
+                      {formErrors.general ? errorIcon : hasNoAllowedTeamsWarning}
                     </span>
                   }
-                  key="general"
+                  key="sharing"
                   forceRender
                 >
+                  <Hideable hidden={this.state.activeTabKey !== "sharing"}>
+                    <ImportSharingComponent
+                      form={form}
+                      datasetId={this.props.datasetId}
+                      hasNoAllowedTeams={_hasNoAllowedTeams}
+                    />
+                  </Hideable>
+                </TabPane>
+
+                <TabPane tab={<span>Metadata</span>} key="general" forceRender>
                   <Hideable hidden={this.state.activeTabKey !== "general"}>
-                    <ImportGeneralComponent form={form} hasNoAllowedTeams={_hasNoAllowedTeams} />
+                    <ImportGeneralComponent form={form} />
                   </Hideable>
                 </TabPane>
